Refactor of hasPermission authorization to use user object and permission object. Added endeavoursecurity typescript generator

+++ src/main/java/org/endeavourhealth/imapi/controllers/WorkflowController.java
@@ -10,18 +10,22 @@ import org.endeavourhealth.imapi.errorhandling.UserNotFoundException;
 import org.endeavourhealth.imapi.filer.TaskFilerException;
 import org.endeavourhealth.imapi.logic.service.SecurityService;
 import org.endeavourhealth.imapi.logic.service.WorkflowService;
-import org.endeavourhealth.imapi.model.security.User;
 import org.endeavourhealth.imapi.model.requests.WorkflowRequest;
 import org.endeavourhealth.imapi.model.responses.WorkflowResponse;
+import org.endeavourhealth.imapi.model.security.Permission;
+import org.endeavourhealth.imapi.model.security.Resource;
+import org.endeavourhealth.imapi.model.security.User;
 import org.endeavourhealth.imapi.model.workflow.BugReport;
 import org.endeavourhealth.imapi.model.workflow.EntityApproval;
 import org.endeavourhealth.imapi.model.workflow.RoleRequest;
 import org.endeavourhealth.imapi.model.workflow.Task;
+import org.endeavourhealth.imapi.model.workflow.roleRequest.UserRole;
 import org.endeavourhealth.imapi.utility.MetricsHelper;
 import org.endeavourhealth.imapi.utility.MetricsTimer;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.context.annotation.RequestScope;
+
+import java.util.List;
 
 @RestController
 @RequestMapping("api/workflow")
@@ -56,20 +60,20 @@ public class WorkflowController {
 
   @Operation(summary = "Update bug report")
   @PostMapping(value = "/updateBugReport")
-  @PreAuthorize("@guard.hasPermission('BUG_REPORT','WRITE')")
   public void updateBugReport(HttpServletRequest request, @RequestBody BugReport bugReport) throws TaskFilerException, UserNotFoundException, JsonProcessingException, UserAuthorisationException {
     try (MetricsTimer t = MetricsHelper.recordTime("API.Workflow.updateBugReport.POST")) {
       log.debug("updateBugReport");
+      securityService.requiresPermission(new Permission(Resource.BUG_REPORT, List.of(UserRole.DEVELOPER), List.of()), request);
       workflowService.updateBugReport(bugReport, request);
     }
   }
 
   @Operation(summary = "Get Tasks by Creator", description = "Retrieve tasks created by the currently authenticated user.")
   @GetMapping(value = "/getTasksByCreatedBy", produces = "application/json")
-  @PreAuthorize("@guard.hasPermission('TASK','READ')")
   public WorkflowResponse getTasksByCreatedBy(HttpServletRequest request, @RequestParam(name = "page", required = false, defaultValue = "1") Integer page, @RequestParam(name = "size", required = false, defaultValue = "25") int size) throws UserNotFoundException, JsonProcessingException {
     try (MetricsTimer t = MetricsHelper.recordTime("API.Workflow.tasksByCreator.GET")) {
       log.debug("getWorkflowsByCreatedBy");
+      securityService.requiresPermission(new Permission(Resource.TASK, List.of(UserRole.DEVELOPER, UserRole.TASK_MANAGER), List.of()), request);
       WorkflowRequest wfRequest = new WorkflowRequest(request);
       if (page != 0) wfRequest.setPage(page);
       if (size != 0) wfRequest.setSize(size);
@@ -79,10 +83,10 @@ public class WorkflowController {
 
   @Operation(summary = "Get Tasks by Assignee", description = "Retrieve tasks assigned to the currently authenticated user.")
   @GetMapping(value = "/getTasksByAssignedTo", produces = "application/json")
-  @PreAuthorize("@guard.hasPermission('TASK','READ')")
   public WorkflowResponse getTasksByAssignedTo(HttpServletRequest request, @RequestParam(name = "page", required = false, defaultValue = "1") Integer page, @RequestParam(name = "size", required = false, defaultValue = "25") Integer size) throws UserNotFoundException, JsonProcessingException {
     try (MetricsTimer t = MetricsHelper.recordTime("API.Workflow.tasksByAssignedTo.GET")) {
       log.debug("getWorkflowsByAssignedTo");
+      securityService.requiresPermission(new Permission(Resource.TASK, List.of(UserRole.DEVELOPER, UserRole.TASK_MANAGER), List.of()), request);
       WorkflowRequest wfRequest = new WorkflowRequest(request);
       if (page != 0) wfRequest.setPage(page);
       if (size != 0) wfRequest.setSize(size);
@@ -92,10 +96,10 @@ public class WorkflowController {
 
   @Operation(summary = "Get Unassigned Tasks", description = "Retrieve tasks that are not assigned to any user.")
   @GetMapping(value = "/getUnassignedTasks", produces = "application/json")
-  @PreAuthorize("@guard.hasPermission('TASK','READ')")
   public WorkflowResponse getUnassignedTasks(HttpServletRequest request, @RequestParam(name = "page", required = false, defaultValue = "1") Integer page, @RequestParam(name = "size", required = false, defaultValue = "25") Integer size) throws UserNotFoundException, JsonProcessingException, UserAuthorisationException {
     try (MetricsTimer t = MetricsHelper.recordTime("API.Workflow.unassignedTasks.GET")) {
       log.debug("getUnassignedTasks");
+      securityService.requiresPermission(new Permission(Resource.TASK, List.of(UserRole.DEVELOPER, UserRole.TASK_MANAGER), List.of()), request);
       WorkflowRequest wfRequest = new WorkflowRequest(request);
       if (page != 0) wfRequest.setPage(page);
       if (size != 0) wfRequest.setSize(size);
@@ -105,19 +109,20 @@ public class WorkflowController {
 
   @Operation(summary = "Get a Task", description = "Fetch a task using its unique ID.")
   @GetMapping(value = "/getTask", produces = "application/json")
-  @PreAuthorize("@guard.hasPermission('TASK','READ')")
-  public Task getTask(@RequestParam(name = "id") String id) throws UserNotFoundException {
+  public Task getTask(HttpServletRequest request, @RequestParam(name = "id") String id) throws UserNotFoundException {
     try (MetricsTimer t = MetricsHelper.recordTime("API.Workfflow.task.GET")) {
       log.debug("getTask");
+      securityService.requiresPermission(new Permission(Resource.TASK, List.of(UserRole.DEVELOPER, UserRole.TASK_MANAGER), List.of()), request);
       return workflowService.getTask(id);
     }
   }
 
   @Operation(summary = "Delete a Task", description = "Delete a task by its unique ID.")
   @DeleteMapping(value = "/deleteTask")
-  @PreAuthorize("@guard.hasPermission('TASK','DELETE')")
-  public void deleteTask(@RequestParam(name = "id") String id) throws TaskFilerException {
+  public void deleteTask(HttpServletRequest request, @RequestParam(name = "id") String id) throws TaskFilerException {
     try (MetricsTimer t = MetricsHelper.recordTime("API.Workflow.task.DELETE")) {
+      log.debug("deleteTask");
+      securityService.requiresPermission(new Permission(Resource.TASK, List.of(UserRole.DEVELOPER, UserRole.TASK_MANAGER), List.of()), request);
       workflowService.deleteTask(id);
     }
   }
@@ -134,40 +139,40 @@ public class WorkflowController {
 
   @Operation(summary = "Get Role Request", description = "Retrieve a role request using its unique ID.")
   @GetMapping(value = "/roleRequest", produces = "application/json")
-  @PreAuthorize("@guard.hasPermission('ROLE_REQUEST','READ')")
   public RoleRequest getRoleRequest(@RequestParam(name = "id") String id, HttpServletRequest request) throws UserNotFoundException, UserAuthorisationException {
     try (MetricsTimer t = MetricsHelper.recordTime("API.Workflow.roleRequest.GET")) {
       log.debug("getRoleRequest");
+      securityService.requiresPermission(new Permission(Resource.ROLE_REQUEST, List.of(UserRole.TASK_MANAGER), List.of()), request);
       return workflowService.getRoleRequest(id);
     }
   }
 
   @Operation(summary = "Update role request", description = "Update a role request workflow task")
   @PostMapping(value = "/updateRoleRequest")
-  @PreAuthorize("@guard.hasPermission('ROLE_REQUEST','WRITE')")
   public void updateRoleRequest(HttpServletRequest request, @RequestBody RoleRequest roleRequest) throws TaskFilerException, UserNotFoundException, JsonProcessingException {
     try (MetricsTimer t = MetricsHelper.recordTime("API.Workflow.updateRoleRequest.POST")) {
       log.debug("updateRoleRequest");
+      securityService.requiresPermission(new Permission(Resource.ROLE_REQUEST, List.of(UserRole.TASK_MANAGER), List.of()), request);
       workflowService.updateRoleRequest(roleRequest, request);
     }
   }
 
   @Operation(summary = "Approve role request")
   @PostMapping(value = "/approveRoleRequest")
-  @PreAuthorize("@guard.hasPermission('ROLE_REQUEST','APPROVE')")
   public void approveRoleRequest(HttpServletRequest request, @RequestBody RoleRequest roleRequest) throws TaskFilerException, UserNotFoundException, JsonProcessingException, UserAuthorisationException {
     try (MetricsTimer t = MetricsHelper.recordTime("API.Workflow.approveRoleRequest.POST")) {
       log.debug("approveRoleRequest");
+      securityService.requiresPermission(new Permission(Resource.ROLE_REQUEST, List.of(UserRole.APPROVER), List.of()), request);
       workflowService.approveRoleRequest(request, roleRequest);
     }
   }
 
   @Operation(summary = "Reject role request")
   @PostMapping(value = "/rejectRoleRequest")
-  @PreAuthorize("@guard.hasPermission('ROLE_REQUEST','APPROVE')")
   public void rejectRoleRequest(HttpServletRequest request, @RequestBody RoleRequest roleRequest) throws TaskFilerException, UserNotFoundException, JsonProcessingException, UserAuthorisationException {
     try (MetricsTimer t = MetricsHelper.recordTime("API.Workflow.rejectRoleRequest.POST")) {
       log.debug("rejectRoleRequest");
+      securityService.requiresPermission(new Permission(Resource.ROLE_REQUEST, List.of(UserRole.TASK_MANAGER), List.of()), request);
       workflowService.rejectRoleRequest(request, roleRequest);
     }
   }
@@ -225,10 +230,10 @@ public class WorkflowController {
 
   @Operation(summary = "Create Entity Approval", description = "Submit an approval request for an entity.")
   @PostMapping(value = "/createEntityApproval")
-  @PreAuthorize("@guard.hasPermission('ENTITY_APPROVAL','WRITE')")
   public void createEntityApproval(HttpServletRequest request, @RequestBody EntityApproval entityApproval) throws TaskFilerException, UserNotFoundException, JsonProcessingException {
     try (MetricsTimer t = MetricsHelper.recordTime("API.Workflow.createEntityApproval.POST")) {
       log.debug("createEntityApproval");
+      securityService.requiresPermission(new Permission(Resource.ENTITY_APPROVAL, List.of(UserRole.EDITOR, UserRole.CREATOR), List.of()), request);
       User user = securityService.getUser(request);
       if (null == entityApproval.getCreatedBy()) entityApproval.setCreatedBy(user.getId());
       workflowService.createEntityApproval(entityApproval);
@@ -237,50 +242,50 @@ public class WorkflowController {
 
   @Operation(summary = "Get entity approval", description = "Get an approval request for an entity by id")
   @GetMapping(value = "/entityApproval")
-  @PreAuthorize("@guard.hasPermission('ENTITY_APPROVAL','READ')")
-  public EntityApproval getEntityApproval(@RequestParam(name = "id") String id) throws UserNotFoundException {
+  public EntityApproval getEntityApproval(HttpServletRequest request, @RequestParam(name = "id") String id) throws UserNotFoundException {
     try (MetricsTimer t = MetricsHelper.recordTime("API.Workflow.entityApproval.GET")) {
       log.debug("getEntityApproval");
+      securityService.requiresPermission(new Permission(Resource.ENTITY_APPROVAL, List.of(UserRole.TASK_MANAGER), List.of()), request);
       return workflowService.getEntityApproval(id);
     }
   }
 
   @Operation(summary = "Update entity approval", description = "Update an approval request for an entity")
   @PostMapping(value = "/updateEntityApproval")
-  @PreAuthorize("@guard.hasPermission('ENTITY_APPROVAL','WRITE')")
   public void updateEntityApproval(HttpServletRequest request, @RequestBody EntityApproval entityApproval) throws TaskFilerException, UserNotFoundException, JsonProcessingException {
     try (MetricsTimer t = MetricsHelper.recordTime("API.Workflow.updateEntityApproval.POST")) {
       log.debug("updateEntityApproval");
+      securityService.requiresPermission(new Permission(Resource.ENTITY_APPROVAL, List.of(UserRole.TASK_MANAGER), List.of()), request);
       workflowService.updateEntityApproval(entityApproval, request);
     }
   }
 
   @Operation(summary = "Approve entity approval")
   @PostMapping(value = "/approveEntityApproval")
-  @PreAuthorize("@guard.hasPermission('ENTITY_APPROVAL','APPROVE')")
   public void approveEntityApproval(HttpServletRequest request, @RequestBody EntityApproval entityApproval) throws TaskFilerException, UserNotFoundException, JsonProcessingException, UserAuthorisationException {
     try (MetricsTimer t = MetricsHelper.recordTime("API.Workflow.approveEntityApproval.POST")) {
       log.debug("approveEntityApproval");
+      securityService.requiresPermission(new Permission(Resource.ENTITY_APPROVAL, List.of(UserRole.APPROVER), List.of()), request);
       workflowService.approveEntityApproval(request, entityApproval);
     }
   }
 
   @Operation(summary = "Reject entity approval")
   @PostMapping(value = "/rejectEntityApproval")
-  @PreAuthorize("@guard.hasPermission('ENTITY_APPROVAL','APPROVE')")
   public void rejectEntityApproval(HttpServletRequest request, @RequestBody EntityApproval entityApproval) throws TaskFilerException, UserNotFoundException, JsonProcessingException, UserAuthorisationException {
     try (MetricsTimer t = MetricsHelper.recordTime("API.Workflow.rejectEntityApproval.POST")) {
       log.debug("rejectEntityApproval");
+      securityService.requiresPermission(new Permission(Resource.ENTITY_APPROVAL, List.of(UserRole.APPROVER), List.of()), request);
       workflowService.rejectEntityApproval(request, entityApproval);
     }
   }
 
   @Operation(summary = "Update Task", description = "Update details of an existing task.")
   @PostMapping(value = "/updateTask")
-  @PreAuthorize("@guard.hasPermission('TASK','WRITE')")
   public void updateTask(HttpServletRequest request, @RequestBody Task task) throws TaskFilerException, UserNotFoundException, JsonProcessingException {
     try (MetricsTimer t = MetricsHelper.recordTime("API.Workflow.updateTask.POST")) {
       log.debug("updateTask");
+      securityService.requiresPermission(new Permission(Resource.TASK, List.of(UserRole.TASK_MANAGER), List.of()), request);
       User user = securityService.getUser(request);
       workflowService.updateTask(task, user.getId());
     }
